requirements.txt
asyncio==3.4.3
black==25.1.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
greenlet==3.1.1
kiwisolver==1.4.8
matplotlib==3.10.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.3.1
PyVISA==1.14.1
PyVISA-py==0.7.2
six==1.17.0
smbus2==0.5.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
websockets==14.2
